subButton cn 설정
